Generics
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@14401 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/model/impl/ListTypeModelImpl.java
@@ -197,7 +197,7 @@ public class ListTypeModelImpl extends BaseModelImpl {
 	}
 
 	public int hashCode() {
-		return (int)getPrimaryKey();
+		return getPrimaryKey();
 	}
 
 	private int _listTypeId;
